Code cleanup, TC refactoring
