+++ tests/convergence/test_convergence_dp_only.py
@@ -1,22 +1,14 @@
 import pytest
 from bgp_convergence_config_b2b import bgp_convergence_config
 
-PRIMARY_ROUTES_NAME = "rx_rr"
-PRIMARY_PORT_NAME = "rx"
-
 
 def test_convergence_dp_only(utils, api, bgp_convergence_config):
     """
     1. set convergence config & start traffic
-    Scenario 1:
-    1. Start traffic
-    2. Withdraw Routes and see events are populated properly
-    Scenario 2:
-    1. Start traffic
-    2. Shutdown primary port and see events are populated properly
+    2. validate convergence metrics for DP only
     """
 
-    # convergence config
+    # convergence config DP only
     bgp_convergence_config.events.dp_events.enable = True
     bgp_convergence_config.events.dp_events.rx_rate_threshold = 90
 
@@ -30,7 +22,6 @@ def test_convergence_dp_only(utils, api, bgp_convergence_config):
     if len(res.warnings) > 0:
         print("Warnings: {}".format(res.warnings))
 
-    # Scenario 1: Route withdraw/Advertise
     # Start traffic
     cs = api.control_state()
     cs.choice = cs.TRAFFIC
@@ -42,7 +33,7 @@ def test_convergence_dp_only(utils, api, bgp_convergence_config):
 
     # Wait for traffic to reach configured line rate
     utils.wait_for(
-        lambda: is_traffic_running(api), "traffic in started state"
+        lambda: utils.is_traffic_running(api), "traffic in started state"
     )
     
     # Port Metrics
@@ -91,20 +82,6 @@ def test_convergence_dp_only(utils, api, bgp_convergence_config):
     # session to another session
     conv_config = api.config()
     api.set_config(conv_config)
-
-
-def is_traffic_running(api):
-    """
-    Returns true if traffic in start state
-    """
-    flow_stats = get_flow_stats(api)
-    return all([int(fs.frames_rx_rate) > 0 for fs in flow_stats])
-
-
-def get_flow_stats(api):
-    request = api.metrics_request()
-    request.convergence.flow_names = []
-    return api.get_metrics(request).flow_metrics
 
 
 if __name__ == "__main__":
